requirements.txt
Django>=4.2
psycopg2-binary
kafka-python
pyzk
gunicorn
whitenoise
